Seventh Draft
ReadFromText method is implemented for GraphSquare class. Fixed bugs involving undo method when shapes are drawn. Application asks for saving when there are changes made without saving to a file. Cursor now works as expected when a shape is selected and while move function is selected. ToolTip is used for displaying function of buttons for better UX.