Fifteenth Draft
Polygon lines are drawn while left clicking. Cursor coordinates and selected button is displayed in bottom left corner in a label. Unit and 2 decimal place is displayed in ruler button. Version and signature is added to About button. DrawCurve function is added. Shortcut for Curvebutton is added. Horizontal and vertical symmetry is added for curve and polygon shapes. Selecting is made easier for polygon and curve shapes. Code is refactored for better readability.