Version 18.0
Cryptology icon is added. Shape's color can be set after it's drawn. Saved files can be encrypted and decrypted. A bug which prevented onformclosing function to work is fixed. Cornerlocation is set to PointF.Empty, after shape is deleted. Encryption key is written in a way to obfuscate for security purposes. Polygon and Curve shapes are moved with mousemove. Background image is located below menustrip and buttons panel.